requirements.txt
phpipam_client
pymongo
mysql-connector-python
redis
protobuf
grpc
grpcio
telegram
requests
